Update to move reusable test code to seperate Forum.Data.Library package
